pipeline/pyproject.toml
[project]
name = "slu-hindex"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scholarly>=1.7.11",
]
